ConDec-526: Improve test set up and data (#158)
* Add all knowledge types by calling KnowledgeType.values() method instead of manually adding them

* Reduce redundant code in class TestSetUpWithIssues

* Remove unused JqlClauseBuilder mock, TimeZoneManager and MockApplicationProperties

Make TestSetUp classes abstract since they do not contain any tests and to avoid "no runnable methods" exception

* Add MockDatabase class to mock the AO databases

* Add MockComponentAccessor class

* Remove SearchService, PluginSettingsFactory, TransactionTemplate, TemplateRenderer, IssueService and ProjectService from ComponentGetter class since it is provided by JIRA's ComponentAccessor class

* Make PLUGIN_KEY a public final static constant in the ComponentGetter class

* Remove inheritance in TestComponentGetter class and from classes in config test package

* Move initialization method to BeforeClass setUp method to create the ComponentAccessor less often

* Remove unused dependencies in pom.xml

* Set "user" attribute in request objects to mock users

* Fix failing test case by providing @NonTransactional annotations that make sure that the AO database tables are not changed by tests

+++ src/test/java/de/uhd/ifi/se/decision/management/jira/mocks/MockIssueLinkManager.java
@@ -41,15 +41,15 @@ public class MockIssueLinkManager implements IssueLinkManager {
 	public List<IssueLink> getInwardLinks(Long arg0) {
 		List<IssueLink> allInwardIssueLink = new ArrayList<>();
 		if (arg0 == 20) {
-			IssueLink link = new MockIssueLink((long) 3);
+			IssueLink link = new MockIssueLink(3);
 			allInwardIssueLink.add(link);
 		}
 		if (arg0 == 30) {
 			for (int i = 0; i < 10; i++) {
-				IssueLink link = new MockIssueLink((long) 3);
-				((MockIssueLink) link).setSequence((long) i);
+				IssueLink link = new MockIssueLink(3);
+				((MockIssueLink) link).setSequence(i);
 				if (i == 9) {
-					((MockIssueLink) link).setSequence((long) 1);
+					((MockIssueLink) link).setSequence(1);
 				}
 				allInwardIssueLink.add(link);
 			}
@@ -105,21 +105,21 @@ public class MockIssueLinkManager implements IssueLinkManager {
 			}
 		}
 		if (arg0 == 20) {
-			IssueLink link = new MockIssueLink((long) 4);
+			IssueLink link = new MockIssueLink(4);
 			allOutwardIssueLink.add(link);
 		}
 		if (arg0 == 30) {
 			for (int i = 0; i < 10; i++) {
-				IssueLink link = new MockIssueLink((long) 3);
+				IssueLink link = new MockIssueLink(3);
 				((MockIssueLink) link).setSequence((long) i + 10);
 				if (i == 9) {
-					((MockIssueLink) link).setSequence((long) 1);
+					((MockIssueLink) link).setSequence(1);
 				}
 				allOutwardIssueLink.add(link);
 			}
 		}
 		if (arg0 == 12) {
-			MockIssueLink link = new MockIssueLink((long) 1);
+			MockIssueLink link = new MockIssueLink(1);
 			allOutwardIssueLink.add(link);
 		}
 		return allOutwardIssueLink;
